refactor(#1382): extract crash_respawn module from daemon/mod.rs
Move handle_crash_respawn, notify_crash, respawn_agent_worker
(170 lines of crash-respawn behavior) from daemon/mod.rs into
dedicated daemon/crash_respawn.rs module.

- DaemonContext + fields promoted to pub(super) for sub-module access
- Remove unused NotifySeverity import from mod.rs
- mod.rs: 2076 → 1904 lines (-172)
